\name{euromap}
\alias{euromap}
\docType{data}
\title{
  Map of Europe
}
\description{
  This dataset is an object of class \code{sfc} from the package sf,
  containing the map of Europe.
}
\usage{data("euromap")}
\format{
  An object of class \code{sf}.
}
\examples{

plot(euromap)

}
\keyword{datasets}
